test: move testharness + e2e/testapp out of internal/ for cross-module use
Go's internal-package visibility rule prevents code outside the module
tree (e.g. github.com/clawvisor/cloud) from importing
github.com/clawvisor/clawvisor/internal/*. That blocked the next step in
the test-infra consolidation, which is for cloud to import the mocks
from this submodule instead of maintaining its own forked copies.

  git mv internal/testharness   → testharness
  git mv internal/e2e/testapp   → e2e/testapp

Plus import-path rewrites across 17 files. Package docs updated to
state these are exported only for cross-module access, with no
API-stability guarantee — they remain intended for clawvisor's and
cloud's test suites only.

No behavior change. All existing testharness + testapp tests pass.
