+++ testharness/github/github.go
@@ -5,7 +5,7 @@ package github
 import (
 	"testing"
 
-	"github.com/clawvisor/clawvisor/internal/testharness/httpmock"
+	"github.com/clawvisor/clawvisor/testharness/httpmock"
 )
 
 type Mock struct {
