+++ testharness/google/google_fix_test.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 	"testing"
 
-	hg "github.com/clawvisor/clawvisor/internal/testharness/google"
+	hg "github.com/clawvisor/clawvisor/testharness/google"
 )
 
 // TestUserinfoReadsAfterTokenConsumed — earlier, /token cleared
