+++ testharness/linear/linear.go
@@ -6,7 +6,7 @@ package linear
 import (
 	"testing"
 
-	"github.com/clawvisor/clawvisor/internal/testharness/httpmock"
+	"github.com/clawvisor/clawvisor/testharness/httpmock"
 )
 
 type Mock struct{ *httpmock.Server }
