+++ testharness/llm/anthropic_live_test.go
@@ -10,7 +10,7 @@ import (
 	"path/filepath"
 	"testing"
 
-	hllm "github.com/clawvisor/clawvisor/internal/testharness/llm"
+	hllm "github.com/clawvisor/clawvisor/testharness/llm"
 )
 
 // TestAnthropicLiveRecordReplay hits the real Anthropic API once to record
