+++ testharness/llm/cassette_test.go
@@ -9,7 +9,7 @@ import (
 	"strings"
 	"testing"
 
-	hllm "github.com/clawvisor/clawvisor/internal/testharness/llm"
+	hllm "github.com/clawvisor/clawvisor/testharness/llm"
 )
 
 // TestRecordThenReplay validates the basic record→replay loop: hit a real
