+++ testharness/microsoft/microsoft.go
@@ -5,7 +5,7 @@ package microsoft
 import (
 	"testing"
 
-	"github.com/clawvisor/clawvisor/internal/testharness/httpmock"
+	"github.com/clawvisor/clawvisor/testharness/httpmock"
 )
 
 type Mock struct{ *httpmock.Server }
